Fix compilation errors and nightly.yml startup_failure
- Fix ObjectPools.kt: use .toInt() for 0x80000000 literal (out of Int range)
- Fix ThermalGuard.kt: use getDeviceTemperatures() API instead of non-existent deviceTemperatures property
- Fix nightly.yml: guard if condition against undefined inputs for non-dispatch events
- Fix nightly.yml: remove inline comment from cron expression

+++ common/src/main/java/com/kurostream/common/pool/ObjectPools.kt
@@ -240,7 +240,7 @@ class SubtitleStyle(
     var primaryColor: Int = 0xFFFFFFFF.toInt(),
     var secondaryColor: Int = 0xFF000000.toInt(),
     var outlineColor: Int = 0xFF000000.toInt(),
-    var backColor: Int = 0x80000000,
+    var backColor: Int = 0x80000000.toInt(),
     var bold: Boolean = false,
     var italic: Boolean = false,
     var underline: Boolean = false,
